refactor(control/v3): extract sessions subpackage and split handlers_sessions.go (A2.2)
Extracted lease consumption, runtime policy, replay, transitions, and playback feedback business logic from package v3 into standalone subpackage backend/internal/control/http/v3/sessions/. Replaced redundant helper methods in handlers_sessions.go with clean delegation to sessions.Service while preserving route table hash (4fb26c9dd199f4e41a072067ce5b93bd invariant) and external API contracts.
