+++ backend/internal/control/http/v3/sessions_runtime_replay.go
@@ -1,43 +1,19 @@
+// Copyright (c) 2025 ManuGH
+// Licensed under the PolyForm Noncommercial License 1.0.0
+// Since v2.0.0, this software is restricted to non-commercial use only.
+
 package v3
 
 import (
-	"strings"
-
+	"github.com/ManuGH/xg2g/internal/control/http/v3/sessions"
 	"github.com/ManuGH/xg2g/internal/control/recordings/runtimepolicy"
 	"github.com/ManuGH/xg2g/internal/domain/session/model"
 )
 
 func buildSessionRuntimePolicyReplay(session *model.SessionRecord) *runtimepolicy.RuntimePolicyReplay {
-	if session == nil {
-		return nil
-	}
-	timeline := loadSessionRuntimeTimeline(session)
-	if len(timeline) == 0 {
-		return nil
-	}
-	first := timeline[0]
-	initial := runtimepolicy.NormalizeSessionLoopState(runtimepolicy.SessionLoopState{
-		CurrentStep:       first.ObservedStep,
-		TargetStep:        first.TargetStep,
-		ConfidenceScore:   first.ConfidenceScore,
-		ConfidenceState:   first.ConfidenceState,
-		CooldownUntil:     first.CooldownUntil,
-		PolicyConstraints: append([]string(nil), first.PolicyConstraints...),
-		Reasons:           append([]string(nil), first.Reasons...),
-	})
-	replay := runtimepolicy.ReplayFromTimeline(runtimepolicy.ReplayMetadata{
-		SessionID:     strings.TrimSpace(session.SessionID),
-		ServiceRef:    strings.TrimSpace(session.ServiceRef),
-		ClientPath:    sessionContextValue(session, model.CtxKeyClientPath),
-		SourceType:    sessionContextValue(session, model.CtxKeySourceType),
-		InitialTarget: first.TargetStep,
-	}, initial, timeline)
-	return &replay
+	return sessions.BuildSessionRuntimePolicyReplay(session)
 }
 
 func sessionContextValue(session *model.SessionRecord, key string) string {
-	if session == nil || session.ContextData == nil {
-		return ""
-	}
-	return strings.TrimSpace(session.ContextData[key])
+	return sessions.SessionContextValue(session, key)
 }
